fix: changed plans are delted anymore

+++ BPBackendDjango/BPBackendDjango/Views/planviews.py
@@ -189,7 +189,20 @@ class CreatePlanView(APIView):
             users = User.objects.filter(plan=req_data['id'])
             for user in users:
                 add_plan_to_user(user.username, plan.id)
-            TrainingSchedule.objects.filter(id=int(req_data['id'])).delete()
+            old_plan = TrainingSchedule.objects.get(id=int(req_data['id']))
+            #check if a doneExercise relates to this plan
+            needed = False
+            for exip in ExerciseInPlan.objects.filter(plan=plan):
+                if DoneExercises.objects.filter(exercise=exip).exists():
+                    needed = True
+            #if yes keep old plan and relate it to new one
+            if needed:
+                old_plan.new_plan = plan
+                old_plan.changed = True
+                old_plan.save(force_update=True)
+            #else delete old plan
+            else:
+                TrainingSchedule.objects.filter(id=int(req_data['id'])).delete()
             data = {
                 'success': True,
                 'description': 'plan was changed',
